[QOL-7939] split CKAN 2.8/2.9 functionality
- separate Pylons from Flask
- separate Paster from Click
